character limit hook
